Calculating time windows: use definition by Hartmann (2013) for both encodings (SMT encoding no longer uses definition by Bofill et al. (2020))
